Gaurav 2/requirements.txt
Django==5.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
django-cors-headers==4.3.1
PyPDF2==3.0.1 